@@ -1,10 +1,10 @@
 
 class Sight{
-  final String name;
-  final double lat, lon;
-  final String url;
-  final String details;
-  final String type;
+  final String? name;
+  final double? lat, lon;
+  final String? url;
+  final String? details;
+  final String? type;
 
   const Sight({this.name, this.lat, this.lon, this.url, this.details, this.type});
   
