@@ -2,7 +2,7 @@ import 'dart:ui';
 
 import 'package:flutter/material.dart';
 
-bool isDarkMode = false;
+bool isDarkMode = true;
 
 Color hexToColor(String code) {
   return new Color(int.parse(code.substring(1, 7), radix: 16) + 0xFF000000);
@@ -17,7 +17,7 @@ Color hexToColor(String code) {
  */
 class BottomNavBar extends StatefulWidget {
   const BottomNavBar({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
