@@ -4,6 +4,7 @@ import 'package:places/mocks.dart';
 import 'package:places/ui/screen/add_sight_screen.dart';
 import 'package:places/ui/screen/filter_screen.dart';
 import 'package:places/ui/screen/res/themes.dart';
+import 'package:places/ui/screen/search_screen.dart';
 import 'package:places/ui/screen/settings_screen.dart';
 import 'package:places/ui/screen/sight_details.dart';
 import 'package:places/ui/screen/sight_list_screen.dart';
@@ -15,7 +16,7 @@ import 'globals.dart' as globals;
 void main() {
   return runApp(ChangeNotifierProvider(
   child: App(),
-  create: (BuildContext context) => ThemeProvider(isDarkMode: false),));
+  create: (BuildContext context) => ThemeProvider(isDarkMode: true),));
 }
 
 class App extends StatefulWidget {
@@ -33,13 +34,14 @@ class _AppState extends State<App> {
       builder: (context, themeProvider, child) {
         return MaterialApp(
         title: 'First App',
+      //  home: SearchScreen(),
        // home: AddSightScreenCategory(setCategory: (value){}),
        // home: AddSightScreenCategory(setCategory: (value){} ),
-        home: AddSightScreen(),
+       // home: AddSightScreen(),
        // home: SettingsScreen(),
         // home: FilterScreen(),
         // home: VisitingScreen(),
-        // home: SightListScreen(title: "Список интересных мест",),
+         home: SightListScreen(title: "Список интересных мест",),
         // home: SightDetails(sight: Sight(name: mocks[0][0], lat: mocks[0][1], lon: mocks[0][2], url: mocks[0][3], details: mocks[0][4], type: mocks[0][5]),),
         //theme: globals.isDarkMode ? darkTheme : lightTheme,
         theme: themeProvider.getTheme,
