@@ -11,19 +11,19 @@ final List<List<String>> mocks = [
     "37.60911009869005",
     "https://bigenc.ru/media/2016/10/27/1235248337/31233.jpg",
     "Здание 1862 года в стиле неоклассицизма, где хранится государственная коллекция книг.",
-    "Библиотека"],
+    "Особое место"],
   ["Белая площадь", 
     "55.77813352604905",
     "37.586219403587094",
     "https://www.belajaploshad.ru/wp-content/uploads/2018/09/white-square.jpg",
     "",
-    "Площадь"],
+    "Особое место"],
   ["Римско-католический Кафедральный собор Непорочного Зачатия Пресвятой Девы Марии", 
     "0.0",
     "0.0",
     "https://i.pinimg.com/originals/ea/01/42/ea0142373b202bb072d402af1821dd6e.jpg",
     "",
-    "Храм"],
+    "Особое место"],
   ["Музей-мастерская Зураба Церетели", 
     "0.0",
     "0.0",
@@ -47,7 +47,7 @@ final List<List<String>> mocks = [
     "39.421353489874484",
     "https://lh5.googleusercontent.com/p/AF1QipNZ7Sy6lw-m7smIFaWdjeHXpFPu7hgo430WCCXo=w408-h306-k-no",
     "Крепость, основанная турками в 1475 году, с земляным валом, сторожевыми башнями и воротами.",
-    "Крепость"],
+    "Особое место"],
     ["Памятник Г.И. Мирошниченко", 
     "47.11780850016745", 
     "39.42098596110988",
@@ -65,6 +65,6 @@ final List<List<String>> mocks = [
     "39.435854337329765",
     "https://lh5.googleusercontent.com/p/AF1QipOu7LNuv6w9PDqMJySwYey9HYhR99YjZIIEQ7AN=w408-h544-k-no",
     "Памятник лётчикам 248-го истребительного полка, погибшим в небе над Азовом, установлен в 1942г на въезде в город, на месте падения первого, сбитого над Азовом немецкого самолета.",
-    "Историческое место"],
+    "Особое место"],
 
 ];
